pyproject.toml
[project]
name = "obp-kd-enose"
version = "0.1.0"
description = "OBP docking prior × e-nose KD pipeline (drift & device transfer robustness)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Jinwon Hwang"}]

# 1단계: 런타임 의존성(가벼운 것만). torch는 4단계에서 추가한다.
dependencies = [
  "PyYAML>=6.0.1",
  "pandas>=2.2",
  "numpy>=1.26"
]

[project.optional-dependencies]
dev = [
  "pre-commit>=3.6",
  "black==24.8.0",
  "ruff>=0.5.6",
  "mypy>=1.10",
  "pytest>=8.0",
  "pytest-cov>=5.0"
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E","F","I","B","UP"]
ignore = ["E203","E266","E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
profile = "black"
